pyproject.toml
[project]
name = "productboardmcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=3.2.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "mypy>=1.20.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=0.25",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
